+++ draftsman/prototypes/logistic_passive_container.py
@@ -22,13 +22,13 @@ from pydantic import ConfigDict
 
 
 class LogisticPassiveContainer(
-    InventoryMixin, 
+    InventoryMixin,
     RequestItemsMixin,
     LogisticModeOfOperationMixin,
     CircuitConditionMixin,
-    ControlBehaviorMixin, 
-    CircuitConnectableMixin, 
-    Entity
+    ControlBehaviorMixin,
+    CircuitConnectableMixin,
+    Entity,
 ):
     """
     A logistics container that provides it's contents to the logistic network
@@ -44,7 +44,10 @@ class LogisticPassiveContainer(
         CircuitConnectableMixin.Format,
         Entity.Format,
     ):
-        class ControlBehavior(LogisticModeOfOperationMixin.ControlFormat, CircuitConditionMixin.ControlFormat):
+        class ControlBehavior(
+            LogisticModeOfOperationMixin.ControlFormat,
+            CircuitConditionMixin.ControlFormat,
+        ):
             pass
 
         control_behavior: Optional[ControlBehavior] = ControlBehavior()