+++ draftsman/signatures.py
@@ -1254,6 +1254,7 @@ class Section(DraftsmanBaseModel):
     @field_validator("filters", mode="before")
     @classmethod
     def normalize_input(cls, value: Any):
+        print("testing...")
         if isinstance(value, list):
             for i, entry in enumerate(value):
                 if isinstance(entry, tuple):
@@ -1267,6 +1268,8 @@ class Section(DraftsmanBaseModel):
                     }
 
         return value
+    
+    model_config = ConfigDict(validate_assignment=True)
 
 
 class QualityFilter(DraftsmanBaseModel):
@@ -1323,99 +1326,97 @@ class ItemRequest(DraftsmanBaseModel):
 
 
 class Sections(DraftsmanBaseModel):
-    class Section(DraftsmanBaseModel):
-        index: uint32
-        filters: Optional[list[SignalFilter]] = []
-        group: Optional[str] = Field(
-            None,
-            description="Name of this particular signal section group.",
-        )
+    # class Section(DraftsmanBaseModel):
+    #     index: uint32
+    #     filters: Optional[list[SignalFilter]] = []
+    #     group: Optional[str] = Field(
+    #         None,
+    #         description="Name of this particular signal section group.",
+    #     )
 
-        def set_signal(
-            self,
-            index: int64,
-            name: Union[str, None],
-            count: int32 = 0,
-            quality: Literal[
-                "normal",
-                "uncommon",
-                "rare",
-                "epic",
-                "legendary",
-                "quality-unknown",
-                "any",
-            ] = "normal",
-            type: Optional[str] = None,
-        ) -> None:
-            try:
-                new_entry = SignalFilter(
-                    index=index,
-                    name=name,
-                    type=type,
-                    quality=quality,
-                    comparator="=",
-                    count=count,
-                )
-                new_entry.index += 1
-            except ValidationError as e:
-                raise DataFormatError(e) from None
+    #     def set_signal(
+    #         self,
+    #         index: int64,
+    #         name: Union[str, None],
+    #         count: int32 = 0,
+    #         quality: Literal[
+    #             "normal",
+    #             "uncommon",
+    #             "rare",
+    #             "epic",
+    #             "legendary",
+    #             "quality-unknown",
+    #             "any",
+    #         ] = "normal",
+    #         type: Optional[str] = None,
+    #     ) -> None:
+    #         try:
+    #             new_entry = SignalFilter(
+    #                 index=index,
+    #                 name=name,
+    #                 type=type,
+    #                 quality=quality,
+    #                 comparator="=",
+    #                 count=count,
+    #             )
+    #             new_entry.index += 1
+    #         except ValidationError as e:
+    #             raise DataFormatError(e) from None
 
-            new_filters = [] if self.filters is None else self.filters
+    #         new_filters = [] if self.filters is None else self.filters
 
-            # Check to see if filters already contains an entry with the same index
-            existing_index = None
-            for i, signal_filter in enumerate(new_filters):
-                if (
-                    index + 1 == signal_filter["index"]
-                ):  # Index already exists in the list
-                    if name is None:  # Delete the entry
-                        del new_filters[i]
-                    else:
-                        new_filters[i] = new_entry
-                    existing_index = i
-                    break
+    #         # Check to see if filters already contains an entry with the same index
+    #         existing_index = None
+    #         for i, signal_filter in enumerate(new_filters):
+    #             if (
+    #                 index + 1 == signal_filter["index"]
+    #             ):  # Index already exists in the list
+    #                 if name is None:  # Delete the entry
+    #                     del new_filters[i]
+    #                 else:
+    #                     new_filters[i] = new_entry
+    #                 existing_index = i
+    #                 break
 
-            if existing_index is None:
-                new_filters.append(new_entry)
+    #         if existing_index is None:
+    #             new_filters.append(new_entry)
 
-        def get_signal(self, index):
-            """
-            Get the :py:data:`.SIGNAL_FILTER` ``dict`` entry at a particular index,
-            if it exists.
+    #     def get_signal(self, index):
+    #         """
+    #         Get the :py:data:`.SIGNAL_FILTER` ``dict`` entry at a particular index,
+    #         if it exists.
 
-            :param index: The index of the signal to analyze.
+    #         :param index: The index of the signal to analyze.
 
-            :returns: A ``dict`` that conforms to :py:data:`.SIGNAL_FILTER`, or
-                ``None`` if nothing was found at that index.
-            """
-            if not self.filters:
-                return None
+    #         :returns: A ``dict`` that conforms to :py:data:`.SIGNAL_FILTER`, or
+    #             ``None`` if nothing was found at that index.
+    #         """
+    #         if not self.filters:
+    #             return None
 
-            return next(
-                (
-                    item
-                    for item in self.filters
-                    if item["index"] == index + 1
-                ),
-                None,
-            )
+    #         return next(
+    #             (item for item in self.filters if item["index"] == index + 1),
+    #             None,
+    #         )
 
-        @field_validator("filters", mode="before")
-        @classmethod
-        def normalize_input(cls, value: Any):
-            if isinstance(value, list):
-                for i, entry in enumerate(value):
-                    if isinstance(entry, tuple):
-                        value[i] = {
-                            "index": i + 1,
-                            "name": entry[0],
-                            "type": next(iter(get_signal_types(entry[0]))),
-                            "comparator": "=",
-                            "count": entry[1],
-                            "max_count": entry[1],
-                        }
+    #     @field_validator("filters", mode="before")
+    #     @classmethod
+    #     def normalize_input(cls, value: Any):
+    #         if isinstance(value, list):
+    #             for i, entry in enumerate(value):
+    #                 if isinstance(entry, tuple):
+    #                     value[i] = {
+    #                         "index": i + 1,
+    #                         "name": entry[0],
+    #                         "type": next(iter(get_signal_types(entry[0]))),
+    #                         "comparator": "=",
+    #                         "count": entry[1],
+    #                         "max_count": entry[1],
+    #                     }
 
-            return value
+    #         return value
+        
+    #     model_config = ConfigDict(validate_assignment=True)
 
     sections: Optional[list[Section]] = Field([], description="""TODO""")
 
@@ -1423,8 +1424,6 @@ class Sections(DraftsmanBaseModel):
         # Custom getitem for this thing specfically
         # return self.sections[key]
         if isinstance(key, str):
-            return next(
-                section for section in self.sections if section.group == key
-            )
+            return next(section for section in self.sections if section.group == key)
         else:
             return self.sections[key]