+++ test/prototypes/test_logistic_request_container.py
@@ -1,6 +1,7 @@
 # test_logistic_request_container.py
 
 from draftsman.constants import LogisticModeOfOperation, ValidationMode
+from draftsman.classes.mixins import RequestFiltersMixin
 from draftsman.entity import (
     LogisticRequestContainer,
     logistic_request_containers,
@@ -70,18 +71,26 @@ class TestLogisticRequestContainer:
                 "requester-chest", position=[0, 0], invalid_keyword="100"
             ).validate().reissue_all()
         with pytest.warns(UnknownKeywordWarning):
-            LogisticRequestContainer(control_behavior={"unused_key": "something"}).validate().reissue_all()
+            LogisticRequestContainer(
+                control_behavior={"unused_key": "something"}
+            ).validate().reissue_all()
         with pytest.warns(UnknownEntityWarning):
-            LogisticRequestContainer("this is not a logistics storage chest").validate().reissue_all()
+            LogisticRequestContainer(
+                "this is not a logistics storage chest"
+            ).validate().reissue_all()
 
         # Errors
         # Raises schema errors when any of the associated data is incorrect
         with pytest.raises(TypeError):
             LogisticRequestContainer("requester-chest", id=25).validate().reissue_all()
         with pytest.raises(TypeError):
-            LogisticRequestContainer("requester-chest", position=TypeError).validate().reissue_all()
+            LogisticRequestContainer(
+                "requester-chest", position=TypeError
+            ).validate().reissue_all()
         with pytest.raises(DataFormatError):
-            LogisticRequestContainer("requester-chest", bar="not even trying").validate().reissue_all()
+            LogisticRequestContainer(
+                "requester-chest", bar="not even trying"
+            ).validate().reissue_all()
         with pytest.raises(DataFormatError):
             LogisticRequestContainer(
                 "requester-chest",
@@ -92,7 +101,9 @@ class TestLogisticRequestContainer:
                 "requester-chest", request_from_buffers="invalid"
             ).validate().reissue_all()
         with pytest.raises(DataFormatError):
-            LogisticRequestContainer(control_behavior="incorrect").validate().reissue_all()
+            LogisticRequestContainer(
+                control_behavior="incorrect"
+            ).validate().reissue_all()
 
     def test_power_and_circuit_flags(self):
         for name in logistic_request_containers:
@@ -102,14 +113,16 @@ class TestLogisticRequestContainer:
             assert container.circuit_connectable == True
             assert container.dual_circuit_connectable == False
 
-    @pytest.mark.skipif("quality" not in mods.mod_list, reason="Quality mod not enabled")
+    @pytest.mark.skipif(
+        "quality" not in mods.mod_list, reason="Quality mod not enabled"
+    )
     def test_quality_inventory_size(self):
         qualities = {
             "normal": 48,
             "uncommon": 62,
             "rare": 76,
             "epic": 91,
-            "legendary": 120
+            "legendary": 120,
         }
         for quality, size in qualities.items():
             chest = LogisticRequestContainer("passive-provider-chest", quality=quality)
@@ -160,12 +173,13 @@ class TestLogisticRequestContainer:
         container = LogisticRequestContainer("requester-chest")
 
         # Shorthand
-        container.request_filters = [
+        section = container.add_section()
+        section.filters = [
             ("iron-ore", 100),
             ("copper-ore", 200),
             ("coal", 300),
         ]
-        assert container.request_filters == [
+        assert container.sections[-1].filters == [
             RequestFilter(index=1, name="iron-ore", count=100),
             RequestFilter(index=2, name="copper-ore", count=200),
             RequestFilter(index=3, name="coal", count=300),
@@ -183,94 +197,95 @@ class TestLogisticRequestContainer:
             RequestFilter(index=3, name="coal", count=300),
         ]
 
-    def test_set_request_filter(self):
-        container = LogisticRequestContainer("requester-chest")
+    # def test_set_request_filter(self): # TODO: reimplement
+    #     container = LogisticRequestContainer("requester-chest")
 
-        container.set_request_filter(0, "iron-ore", 100)
-        container.set_request_filter(1, "copper-ore", 200)
-        assert container.request_filters == [
-            RequestFilter(index=1, name="iron-ore", count=100),
-            RequestFilter(index=2, name="copper-ore", count=200),
-        ]
+    #     container.set_request_filter(0, "iron-ore", 100)
+    #     container.set_request_filter(1, "copper-ore", 200)
+    #     assert container.request_filters == [
+    #         RequestFilter(index=1, name="iron-ore", count=100),
+    #         RequestFilter(index=2, name="copper-ore", count=200),
+    #     ]
 
-        # Replace existing
-        container.set_request_filter(1, "wooden-chest", 123)
-        assert container.request_filters == [
-            RequestFilter(index=1, name="iron-ore", count=100),
-            RequestFilter(index=2, name="wooden-chest", count=123),
-        ]
+    #     # Replace existing
+    #     container.set_request_filter(1, "wooden-chest", 123)
+    #     assert container.request_filters == [
+    #         RequestFilter(index=1, name="iron-ore", count=100),
+    #         RequestFilter(index=2, name="wooden-chest", count=123),
+    #     ]
 
-        # Omitted count
-        container.set_request_filter(0, "steel-chest")
-        assert container.request_filters == [
-            RequestFilter(index=1, name="steel-chest", count=None),
-            RequestFilter(index=2, name="wooden-chest", count=123),
-        ]
+    #     # Omitted count
+    #     container.set_request_filter(0, "steel-chest")
+    #     assert container.request_filters == [
+    #         RequestFilter(index=1, name="steel-chest", count=None),
+    #         RequestFilter(index=2, name="wooden-chest", count=123),
+    #     ]
 
-        # Delete existing
-        container.set_request_filter(1, None)
-        assert container.request_filters == [
-            RequestFilter(index=1, name="steel-chest", count=None),
-        ]
+    #     # Delete existing
+    #     container.set_request_filter(1, None)
+    #     assert container.request_filters == [
+    #         RequestFilter(index=1, name="steel-chest", count=None),
+    #     ]
 
-        # None case
-        container.request_filters = None
-        assert container.request_filters == None
+    #     # None case
+    #     container.request_filters = None
+    #     assert container.request_filters == None
 
-        # Create from None
-        container.set_request_filter(0, "copper-ore", 200)
-        assert container.request_filters == [
-            RequestFilter(index=1, name="copper-ore", count=200),
-        ]
+    #     # Create from None
+    #     container.set_request_filter(0, "copper-ore", 200)
+    #     assert container.request_filters == [
+    #         RequestFilter(index=1, name="copper-ore", count=200),
+    #     ]
 
-        with pytest.raises(DataFormatError):
-            container.set_request_filter("incorrect", "incorrect")
+    #     with pytest.raises(DataFormatError):
+    #         container.set_request_filter("incorrect", "incorrect")
 
-        # Unknown item
-        # No warning with minimum
-        container.validate_assignment = "minimum"
-        assert container.validate_assignment == ValidationMode.MINIMUM
-        container.set_request_filter(0, "who-knows?", 100)
-        assert container.request_filters == [
-            RequestFilter(index=1, name="who-knows?", count=100),
-        ]
+    #     # Unknown item
+    #     # No warning with minimum
+    #     container.validate_assignment = "minimum"
+    #     assert container.validate_assignment == ValidationMode.MINIMUM
+    #     container.set_request_filter(0, "who-knows?", 100)
+    #     assert container.request_filters == [
+    #         RequestFilter(index=1, name="who-knows?", count=100),
+    #     ]
 
-        container.validate_assignment = ValidationMode.STRICT
-        with pytest.warns(UnknownItemWarning):
-            container.set_request_filter(0, "who-knows?", 100)
+    #     container.validate_assignment = ValidationMode.STRICT
+    #     with pytest.warns(UnknownItemWarning):
+    #         container.set_request_filter(0, "who-knows?", 100)
 
-    def test_set_request_filters(self):
-        container = LogisticRequestContainer("requester-chest")
+    # @pytest.mark.skipif(__factorio_version_info < (2,), "1.0")
+    # def test_set_request_filters(self): # TODO: reimplement
+    #     container = LogisticRequestContainer("requester-chest")
 
-        # Shorthand
-        container.set_request_filters(
-            [("iron-ore", 100), ("copper-ore", 200), ("coal", 300)]
-        )
-        assert container.request_filters == [
-            RequestFilter(index=1, name="iron-ore", count=100),
-            RequestFilter(index=2, name="copper-ore", count=200),
-            RequestFilter(index=3, name="coal", count=300),
-        ]
+    #     # Shorthand
+    #     container.set_request_filters(
+    #         [("iron-ore", 100), ("copper-ore", 200), ("coal", 300)]
+    #     )
+    #     assert container.request_filters == [
+    #         RequestFilter(index=1, name="iron-ore", count=100),
+    #         RequestFilter(index=2, name="copper-ore", count=200),
+    #         RequestFilter(index=3, name="coal", count=300),
+    #     ]
 
-        # Longhand
-        container.set_request_filters(
-            [
-                {"index": 1, "name": "iron-ore", "count": 100},
-                {"index": 2, "name": "copper-ore", "count": 200},
-                {"index": 3, "name": "coal", "count": 300},
-            ]
-        )
-        assert container.request_filters == [
-            RequestFilter(index=1, name="iron-ore", count=100),
-            RequestFilter(index=2, name="copper-ore", count=200),
-            RequestFilter(index=3, name="coal", count=300),
-        ]
+    #     # Longhand
+    #     container.set_request_filters(
+    #         [
+    #             {"index": 1, "name": "iron-ore", "count": 100},
+    #             {"index": 2, "name": "copper-ore", "count": 200},
+    #             {"index": 3, "name": "coal", "count": 300},
+    #         ]
+    #     )
+    #     assert container.request_filters == [
+    #         RequestFilter(index=1, name="iron-ore", count=100),
+    #         RequestFilter(index=2, name="copper-ore", count=200),
+    #         RequestFilter(index=3, name="coal", count=300),
+    #     ]
 
-        # Ensure error in all circumstances
-        container.validate_assignment = "none"
-        assert container.validate_assignment == ValidationMode.NONE
-        with pytest.raises(DataFormatError):
-            container.set_request_filters("incorrect")
+    #     # Ensure error in all circumstances
+    #     container.validate_assignment = "none"
+    #     assert container.validate_assignment == ValidationMode.NONE
+    #     with pytest.raises(DataFormatError):
+    #         container.set_request_filters("incorrect")
 
     def test_request_from_buffers(self):
         container = LogisticRequestContainer("requester-chest")
@@ -285,7 +300,7 @@ class TestLogisticRequestContainer:
         assert container.to_dict() == {
             "name": "requester-chest",
             "position": {"x": 0.5, "y": 0.5},
-            "request_from_buffers": True,
+            "request_filters": {"request_from_buffers": True}
         }
 
         with pytest.raises(DataFormatError):
@@ -299,7 +314,7 @@ class TestLogisticRequestContainer:
         assert container.to_dict() == {
             "name": "requester-chest",
             "position": {"x": 0.5, "y": 0.5},
-            "request_from_buffers": "incorrect",
+            "request_filters": {"request_from_buffers": "incorrect"}
         }
 
     def test_mergable_with(self):
@@ -324,7 +339,16 @@ class TestLogisticRequestContainer:
         container2 = LogisticRequestContainer(
             "requester-chest",
             bar=10,
-            request_filters=[{"name": "utility-science-pack", "index": 1, "count": 10}],
+            request_filters={
+                "sections": [
+                    {
+                        "index": 1,
+                        "filters": [
+                            {"name": "utility-science-pack", "index": 1, "count": 10}
+                        ],
+                    }
+                ]
+            },
             tags={"some": "stuff"},
         )
 
@@ -332,9 +356,16 @@ class TestLogisticRequestContainer:
         del container2
 
         assert container1.bar == 10
-        assert container1.request_filters == [
-            RequestFilter(**{"name": "utility-science-pack", "index": 1, "count": 10})
-        ]
+        assert container1.request_filters == LogisticRequestContainer.Format.LogisticsRequestFilters(
+            sections=[
+                {
+                    "index": 1,
+                    "filters": [
+                        {"name": "utility-science-pack", "index": 1, "count": 10}
+                    ],
+                }
+            ]
+        )
         assert container1.tags == {"some": "stuff"}
 
     def test_eq(self):